feat(timezone): add timezone to manager periods
